+++ libft/ft_strnstr.c
@@ -6,26 +6,26 @@
 /*   By: elaachac <elaachac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 17:35:27 by elaachac          #+#    #+#             */
-/*   Updated: 2019/11/12 13:38:52 by elaachac         ###   ########.fr       */
+/*   Updated: 2019/11/18 17:08:28 by elaachac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char    *ft_strnstr(const char *haystack, const char *needle, size_t len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-    int i;
-    int k;
-    int j;
+	size_t i;
+	size_t k;
+	size_t j;
 
-    i = 0;
-    j = 0;
-    k = 0;
-    if(!(needle))
-        return ((char *)haystack);
-    while(i <= len)
-    {
-        k = i;
+	i = 0;
+	j = 0;
+	k = 0;
+	if (!(needle))
+		return ((char *)haystack);
+	while (i <= len)
+	{
+		k = i;
 		while (haystack[k] == needle[j])
 		{
 			k++;
@@ -36,6 +36,6 @@ char    *ft_strnstr(const char *haystack, const char *needle, size_t len)
 		k = 0;
 		j = 0;
 		i++;
-    }
-    return (NULL);
+	}
+	return (NULL);
 }
